Fixing the problem with the test
I hope.
